% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssj_awt.R
\name{auto_ssj_atw}
\alias{auto_ssj_atw}
\title{Automatic selection of the optimal smoothing spline function with jumps
and fixed smoothing parameter}
\usage{
auto_ssj_atw(
  y,
  x,
  lambda,
  grid = seq(0, sd(y, na.rm = TRUE)/mean(diff(x)) * 10, sd(y, na.rm =
    TRUE)/mean(diff(x))/10),
  ic = c("bic", "hq", "aic", "aicc"),
  max_iter = 100,
  tol = 1e-04,
  learning_rate = 0.1
)
}
\arguments{
\item{y}{vector with the y (missing values allowed).}

\item{x}{vector with the x (no missing values allowed).}

\item{lambda}{smoothing constant;}

\item{grid}{numeric vector containing the grid for the argument \code{maxsum}
of the \code{ssj_atw} function;}

\item{ic}{string with information criterion for the choice: the default is
"bic" (simulations show this is the best choice), but also "hq", "aic" and "aicc"
are available;}

\item{max_iter}{maximum number of iterations in ssj_atw (default 100).}

\item{tol}{tolerance for convergence of ssj_atw (default 1e-4).}

\item{learning_rate}{learning rate of ssj_atw (default 0.1).}
}
\value{
The ouput of the \code{ssj_awt} function corresponding to the best
choice according to the selected information criterion.
}
\description{
The regularization parameter for the HP filter with jumps is the
maximal sum of additional times. This value
has to be passed to the \code{ssj_atw} function. The function \code{auto_ssj_atw}
runs \code{ssj_atw} on a grid of regularization constants and returns the
output of \code{ssj_atw} according to the chosen information criterion.
Each grid point is warm-started from the solution at the previous point.
The grid search stops early when the selected IC increases for three
consecutive grid values.
}
\examples{
plot(faithful$eruptions, faithful$waiting)
of_jum <- auto_ssj_atw(y = faithful$waiting,
                       x = faithful$eruptions,
                       lambda = 1)
lines(x = sort(faithful$eruptions),
      y = of_jum$smoothed_level, col = "red", lwd = 2)

}
