% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssj_fix.R
\name{auto_ssj_fix}
\alias{auto_ssj_fix}
\title{Automatic selection of the optimal smoothing spline function with jumps
and fixed smoothing parameter}
\usage{
auto_ssj_fix(
  y,
  x,
  lambda,
  grid = seq(0, sd(y) * 10, sd(y)/10),
  ic = c("bic", "hq", "aic", "aicc"),
  edf = TRUE,
  last_delta = 1
)
}
\arguments{
\item{y}{vector with the y (missing values allowed).}

\item{x}{vector with the x (no missing values allowed).}

\item{lambda}{smoothing constant;}

\item{grid}{numeric vector containing the grid for the argument \code{maxsum}
of the \code{ssj_fix} function;}

\item{ic}{string with information criterion for the choice: the default is
"bic" (simulations show this is the best choice), but also "hq", "aic" and "aicc"
are available;}

\item{edf}{logical scalar: TRUE (default) if the number of degrees of freedom
should be computed as "effective degrees of freedom" (Efron, 1986) as opposed
to a more traditional way (although not supported by theory) when FALSE.}
}
\value{
The ouput of the \code{ssj_fix} function corresponding to the best
choice according to the selected information criterion.
}
\description{
The regularization parameter for the HP filter with jumps is the
maximal sum of standard deviations added to the static std.dev. This value
has to be passed to the \code{ssj_fix} function. The function \code{auto_ssj_fix}
runs \code{ssj_fix} on a grid of regularization constants and returns the
output of \code{ssj_fix} according to the chosen information criterion.
}
\examples{
plot(faithful$eruptions, faithful$waiting)
of_jum <- auto_ssj_fix(y = faithful$waiting,
                       x = faithful$eruptions,
                       lambda = 1)
lines(x = sort(faithful$eruptions),
      y = of_jum$smoothed_level, col = "red", lwd = 2)

}
