% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.R
\name{BIC.hpj}
\alias{BIC.hpj}
\title{BIC method for the class hpj}
\usage{
\method{BIC}{hpj}(object, ...)
}
\arguments{
\item{object}{an object of class hpj;}

\item{...}{additional objects of class hpj;}
}
\description{
BIC method for the class hpj
}
