% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employed_IT}
\alias{employed_IT}
\title{Dataset: employed_IT}
\format{
A multivariate time series with 126 rows and 10 columns:
\describe{
  \item{Y15.24}{Thousand of employed in the age class 15-24}
  \item{Y15.64}{Thousand of employed in the age class 15-64}
  \item{Y20.64}{Thousand of employed in the age class 20-64}
  \item{Y25.54}{Thousand of employed in the age class 25-54}
  \item{Y55.64}{Thousand of employed in the age class 55-64}
  \item{Y15.29}{Thousand of employed in the age class 15-29}
  \item{Y15.74}{Thousand of employed in the age class 15-74}
  \item{Y25.29}{Thousand of employed in the age class 25-29}
  \item{Y29.54}{Thousand of employed in the age class 29-54}
  \item{Y29.74}{Thousand of employed in the age class 29-74}
}
}
\source{
Eurostat
}
\usage{
employed_IT
}
\description{
A dataset containing the quarterly time series of employed
people by age class in Italy in the time span 1992Q4-2024Q1.
}
\keyword{datasets}
