% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.R
\name{logLik.hpj}
\alias{logLik.hpj}
\title{logLik method for the class hpj}
\usage{
\method{logLik}{hpj}(object)
}
\arguments{
\item{object}{an object of class hpj;}
}
\description{
logLik method for the class hpj
}
