% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(y, x)
}
\arguments{
\item{y}{vector}

\item{x}{vector}
}
\value{
The scalar mean squared difference/error.
}
\description{
It computes the mean squared difference between the elements
of the vectors x and y.
}
