% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.R
\name{plot.hpj}
\alias{plot.hpj}
\title{Plot method for the class hpj}
\usage{
\method{plot}{hpj}(
  x,
  prob = NULL,
  show_breaks = TRUE,
  main = "original + filter",
  use_ggplot = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class hpj;}

\item{prob}{coverage probability for the confidence interval of the filter;}

\item{show_breaks}{logical, if \code{TRUE} vertical lines are drawn at the jumps;}

\item{main}{title of the plot;}

\item{use_ggplot}{logical, if \code{TRUE} the plot is done with ggplot2;}

\item{...}{additional arguments passed to the plot function (if no ggplot2 is used).}
}
\value{
If \code{use_ggplot} is \code{TRUE} a ggplot object is returned, otherwise a base plot is shown.
}
\description{
Plot method for the class hpj
}
