% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssd.R
\name{predict_cssd_potts}
\alias{predict_cssd_potts}
\title{Predict from a fitted CSSD on new x locations}
\usage{
predict_cssd_potts(fit, xnew, y_orig, x_orig, delta = 1)
}
\arguments{
\item{fit}{object returned by \code{ssj_cssd_potts}.}

\item{xnew}{numeric vector of query locations.}

\item{y_orig, x_orig}{original data vectors (needed for re-fitting Hermite
control points on each segment).}

\item{delta}{positive scalar or vector: noise std devs (same as used to
fit \code{fit}).}
}
\value{
Numeric vector of fitted values at \code{xnew}.
}
\description{
Evaluates the piecewise cubic spline returned by \code{\link{ssj_cssd_potts}}
at arbitrary query points.
}
