% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.R
\name{print.hpj}
\alias{print.hpj}
\title{Print method for the class hpj}
\usage{
\method{print}{hpj}(x, ...)
}
\arguments{
\item{x}{an object of class hpj;}

\item{...}{not used: for consistency with generic function.}
}
\description{
Print method for the class hpj
}
