% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssj_fix.R
\name{ssj_fix}
\alias{ssj_fix}
\title{Smoothing splines with with automatic jumps detection and
fixed smoothing parameter}
\usage{
ssj_fix(
  y,
  x,
  lambda,
  maxsum = sd(y, na.rm = TRUE)/mean(diff(x)),
  edf = TRUE,
  parinit = NULL,
  last_delta = 1
)
}
\arguments{
\item{y}{vector with the y (missing values allowed).}

\item{x}{vector with the x (no missing values allowed).}

\item{lambda}{a numeric scalar with the smoothing parameter.}

\item{maxsum}{maximum sum of additional standard deviations.}

\item{edf}{boolean if TRUE computes effective degrees of freedom otherwise computes
the number of degrees of freedom in the LASSO-regression way.}

\item{parinit}{either NULL or vector of 1+n parameters with the starting values for the
optimizer; the order of the parameters is the constant std. dev and
n additional std. devs.}

\item{last_delta}{numeric scalar with the difference \eqn{x_{n+1}-x_{n}}; it is
not relevant for most applications and defaults to 1.}
}
\value{
list with the following slots:
\itemize{
 \item opt: the output of the optimization function (nloptr)
 \item nobs: number of observations
 \item df: model's degrees of freedom
 \item loglik: value of the log-likelihood at maximum
 \item ic: vector of information criteria (aic, aicc, bic, hq)
 \item smoothed_level: vector with smoothing splines with jumps
 \item var_smoothed_level: vector with variances of the smoothing splines
 \item x: the ordered x variable
}
}
\description{
This is the lower-level function for the smoothing splines with jumps and fixed
smoothing parameter.
}
\examples{
plot(faithful$eruptions, faithful$waiting)
of_jum <- ssj_fix(y = faithful$waiting,
                  x = faithful$eruptions,
                  lambda = 1,
                  maxsum = 150)
lines(x = sort(faithful$eruptions),
      y = of_jum$smoothed_level, col = "red", lwd = 2)
      
}
