% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.R
\name{logLik.hpj}
\alias{logLik.hpj}
\title{logLik method for the class hpj}
\usage{
\method{logLik}{hpj}(object, ...)
}
\arguments{
\item{object}{an object of class hpj;}

\item{...}{not used: for consistency with generic function.}
}
\value{
An object of logLik class with the log-likelihood value and two
attributes: df, the number of degrees of freedom, and nobs, the number of
observations.
}
\description{
logLik method for the class hpj
}
