% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.R
\name{nobs.hpj}
\alias{nobs.hpj}
\title{nobs method for the class hpj}
\usage{
\method{nobs}{hpj}(object, ...)
}
\arguments{
\item{object}{an object of class hpj;}

\item{...}{not used: for consistency with generic function.}
}
\description{
nobs method for the class hpj
}
